Added Enemies
Added enemies class/constructor, behaviors/AI, and definitions. Added a single placeholder enemy which uses the same assets as the skeleton warrior.

Added the ability for enemies (and skeletons once added) to target the closest attack target.

Still need to add some actual enemies, and then extend the behavior of them and the skeletons to actually have combat.

Also modified the player and skeletons' movespeed to be a bit slower (as the game progresses, movespeed can increase)