Entities can now select an attack target
For enemies, they will query for attack targets every 1s, then choose the closest target that is in LOS (still have to add objects that can block LOS). From there they will move toward that target until they are in attacking range.

For Players/skeletons, the player will query for any enemy that is on screen every 1s, then skeletons access that table of enemies found by the player and select the closest enemy within LOS and move toward it. Skeletons will still move to the player if they have strayed too far.

Still have to implement attacking. Ranged attacking is easy; just launch a projectile that will damage any target it hits. Melee attacks will be more difficult however. I'll need to have the attacker query the world in an area in front of them (the exact point will be determined by the angle they were moving in) but only during the attack animation frames when the attack should be happening. i.e. on the sword swing, not anticipation. Then we have to ensure that the same target can't be damaged again by that source for some cooldown period, so I'll need to keep track of entities and their recent damage sources.