Melee entities can now knockback and kill
Melee combat (for lunging based melee entities) is now essentially finished, with them being able to damage, knockback, and kill their attack targets, removing them from the game.

Skeletons and enemies are now assigned functions defined in behaviors.methods. This was done to make things a bit more 'object oriented', even though it only really changes the syntax of the function calls. This also means that certain entities will receive methods that they'll never call, but it shouldn't be a huge issue for this game.

Also moved skeleton's and enemy's update and draw functions into their respective behaviors definitions where they'll be assigned to newly constructed entities as methods like changeState and updateAnimation. This was done to prevent redundant memory allocation as the update and draw functions across all like entities are the same.

Also updated the model for the skeletonArcher's bow and arrow as well as the attack animation.

Also increased the zoom from 2 to 3 as 2 was a bit too far. 3 seems a bit too close for what I want but the 'fov' can be increased with a different display resolution.