Fixed attack speed bug
Fixed a bug that caused entities to not attack at the rate defined by their attack speed. This was because after an attack was performed, the entity would go into the 'idle' and clear it's moveTarget (which also serves as the attack target). Then it would need to wait until it was ready to query for a new attack target, which is done at a rate of 1s before being able to re-enter combat. Basically the queryAttackTargetRate and the attackSpeed conflicted with each other, making the entity only seldom attack at it's attackSpeed.

To fix this, if the entity is not ready to attack, it will go into the 'idle' state but will remember the attackTarget and will wait for either its 'canAttack' cooldown to finish or the target moves out of attack range to once again start moving into position for attack.

Also removed stale references to Enemies and Skeletons files.