Split entities' move state into multiple states
The move() state for skeletons and enemies are now split into more dedicated sub-states: moveToTarget() and moveToPlayer(). This should help with readability, simplicity, and is a better standard. This is also in preparation for adding in the moveToLocation() where an entity reposition to a general location. This can be useful for idle wandering or even fleeing.

Fixed a bug causing the angle of knockback to be incorrect for projectiles. This was caused because projectiles ignored the 'enemy' collision class, which often caused a projectile's enter() function to be called when the projectile is already past the (x,y) of an enemy. This was fixed by splitting the 'projectiles' collision class into 'allyProjectiles' and 'enemyProjectiles'.

Updated the behavior of the spark projectile to make it jitter around to look more like electricity.

Updated the way projectiles' update and draw function are passed to new instances. Now, unless defined, new instances with just get the ['base'] update and draw functions.