Optimized for loops
All loops which unnecessarily used pairs() were replaced by regular counting For loops as this is much more efficient.

The only unnecessary pairs() use that  I could not replace was in the TimerState as it was able to resize the table being iterated on during iterations. I was still able to just use ipairs() which is supposedly faster than pairs() though which is better than nothing.