Updated projectiles, optimized callbacks
Updated projectiles class to have separate projectile travelling behavior and projectile on hit (collision with enemies) behavior. This was done because many projectiles will have the 'base' travel behavior, but a different onHit behvaior such as explode, or apply status.

Also added the 'bone' projectile which is launched by the player.

Increased the render distance to 800x600. Projectiles that exceed a distance from the camera of 400 will automatically destroy themselves in addition to their usual destroy on timeup.

Optimized the code in TimerState to write once and reuse timer update functions upon creating a timer instance. Because so many entities will be in the game each with many timers constantly being added, this will help save on memory. There's still room to improve by reducing redundant function allocation upon instantiation.