Projectiles migrated
Projectiles are now the same as they were when using Windfield, except they now use bump and are much more efficient.

Collision behavior is even more realistic as the knockback angle is calculated from the center points of the projectile and the target at the moment of collision (before it just used the angle the projectile was moving in last). This gives the illusion of using 'circle colliders' for each entity.

Still need to implement sparks bouncing off solid walls, which I'll do next after creating a test arena such solid walls.

Also changed the way entities and shadows are drawn so that they rotate around their center point correctly, and so that they line up perfectly with the center point of their collider.

Defined the collision filters for 'ally', 'enemy', 'allyProjectile', and 'enemyProjectile'.

Updated the way cooldowns are defined and work. Cooldowns are now tables with a boolean flag, a cooldown period, and callback functions for releaseCooldown and setOnCooldown (which are generated by timerState). This done to ensure that absolutely no functions are being defined when using timerState:after(), which is much better on memory. It also has the added bonus of making cooldown management more declarative. I.e:
if player.canAttack.flag then
    player:attack()
    player.canAttack.setOnCooldown()
end