Added pumpkin, golem, and spider
Added entity sprite sheets along with their definitions and AI.

Removed the render distance from Objects.lua. Now, all entities will be drawn and updated no matter where they are in the level, which shouldn't affect performance since levels will only be small to medium arenas, no bigger than a few screens. This was also done so that no matter where a skeleton or an enemy is, they'll be able to find the player.

Entities now correctly push other entities out of their collision box with a force the speed of the entity being pushed, not the speed of the pusher. Now, a spider won't push away a golem faster than it can naturally move.

Increased the 'player's center offset point' so that skeletons are a bit more spread out when following the player.

Added knockback resistance (kbResistance) which is used to reduce the amount of knockback done to an entity when it takes damage. The default value is 0 and the max value is 100. Currently only the golem has any kbResistance.