Added more entities, modified some systems
Added bat, zombie, and possessedArcher. Added just the spriteSheet for slimeMatron.

Assigning AI to entities now reuses AI 'bundles' to reduce verbose code. Example: all skeletonMages now just use the skeletonArcher AI

Entities can now have alternate spriteSheets which will be chosen at random upon spawning the entity. So far, only the zombie uses this to spawn with one of three different sprites.

Added the darkArrow projectile which is the arrow used by the possessedArcher. Modified the projectiles class to allow for different projectile collisionClasses. (Before it was just the allyProjectile)

Shadows.lua has been modified to make it easier to define shadows and to let certain shadows reuse already defined sprites. I tried to make it so the shadows could just parse the entitiesDefinitions, but that lead to some non deterministic errors that I believe stem from the load order of the player, entities, and shadows. So essentially, now the shadow names are defined and then that table is parsed to create the table of sprites.