Added entities, upgraded projectile systems
Added possessedKnight, undeadMiner, ent, and headlessHorseman. HeadlessHorseman has exploding projectile is fully functioning, but still needs a particle system or explosion graphics to be complete.

Projectile based enemies need to have their AI expanded so that they move around/relocate in between shots, instead of only moving into range and then idling until their attack is ready.

'explosion' onHit and 'apple' update projectile functions have been added. Explosion is self explanatory, 'apple' is identical to 'base' except that it doesn't rotate to point toward it direction of travel, though it still faces left or right correctly.

The spark projectile now bounces off of solids correctly.

Projectile systems have been modified to let all collisions be handled by the projectile's onHit function. Previously some checking for targets or solid walls were done in the update, then passed to the onHit to deal with any targets that need to be damaged. Now, all collisions are simply passed to the onHit function. This was done as it made more sense, was better coding practice, and was necessary for the 'bouncing' of the spark projectile.