Fixed table.remove() related query bugs
Using table.remove() on a table as it's being iterated upon caused bugs for querying related actions such as getting attack targets and move targets, regardless of iterating using simple integer loops or ipairs(). This is fixed by building another table of 'valid elements' instead of removing invalid elements from the current table.

objects.lua still uses table.remove() on the table it's iterating upon with ipairs() however it doesn't seem to give a 'nill' error. I suspect this is because after iterating over every element, it won't do anything more referencing of the table's elements.

Entities now also have three different world filters: collisionFilter is the filter used to determine collision behavior in updatePosition(), moveFilter is used to query the world for LOS to a potential moveTarget, and losFilter is used to query the world for LOS to a potential attack target. Each of these have defaults, but can be overwritten in the entity's definition.