Implemented grid based terrain placing algorithm
Level manager now divides the map's spawn area into a grid of 16x16 tiles which it uses as a spatial database so that it can place terrain objects correctly (no overlap with other terrain and no terrain being partially out of bounds).

Terrain objects also have a 1-tile wide border surrounding them to ensure there's some space between all terrain objects to move through.

the tombstone is now larger to be easier to hit.

The objects class is now responsible for destroying the colliders of objects when they die/are destroyed. This also means that any physical object that doesn't get added to the objects table (like invisible level boundaries) will need to be manually destroyed when appropriate.

As a safeguard against freezes due to collision overload, if the timestep between frames becomes too small, the level manager will clear the objects table, destroying all objects in the process other than the player. This is only a test solution, I'll need to improve upon this further to ensure no one experiences a collision based game freeze.