Level manager can now spawn waves of enemies
Level manager uses the grid class to spawn enemies in open space according to the specified level definition. After enemies of a certain wave are killed, the next wave of enemies is spawned. Once the final wave of the level is completed, the level will end and the next level will begin (yet to be added).