Enemies waves will no longer spawn near the player
Just before a wave of enemies are spawned, the level grid updates which tiles are currently taken by the player in order to keep any enemies from spawning too close to them.

Level definitions will now have 'maxEnemies' which will be used to restrict how many enemies can be in a level. This is really just to keep summoner entities from endlessly spawning minions until the player is overwhelmed or the game freezes. The only current exception is the spiderEgg, which can only summon minions once.

 Added some extra terrain and decorations.

Finished and improved the slimeMatron animations.