Added ghost and poltergeist. Added level exits
Ghosts and poltergeists are entities which have the ability to teleport after each attack. Ghosts are melee while poltergeists are ranged.

Added exits class which constructs a level exit. Once the player clears all enemy waves in a level, the levelManager will either spawn the exit to a specified position or use the grid to spawn it in a random location. In the future, I plan to delay this spawn until after the camera pans over to it. This will be after the camPanState or equivalent is added.

Currently only the 'hole in the wall' exits for the swamp and caves and the dungeon stairs are added, though I might add some more variation later.