Added fading in/out and camera panning
Added game states for fading in/out and for camera panning.

fadeState allows for fading in/out, callback functions for when the fade is complete, and hold time to specify a period of time to hold the state after the fade is complete. fadeBoth() can also be used to fade out, call a callback function, hold for a period of time, then fade back in with no callback or hold.

panState allows for camera panning to one or more panObjects. Each object will have a target and can optionally also have a callback function which will be called when the pan is complete and a hold time which will hold the camera on the target for a period of time. This allows for chaining camera pans which is what I plan to do; once a level is complete, we'll pan to the reward chest, spawn it, then pan over to the level exit, spawn that, then pan back to the player. In the future I might expand panState to also allow changing the camera smoother which can affect the speed of the camera pan, but for now it's good.

 Skeletons will now hold over between levels. So if the player completed a level with one of each skeleton, they will all spawn with the player once the next level begins. This is done by spawning new instances of them, which means their health will be fully restored between levels. I can change this in the future, but for now I'm okay with this.

Objects.lua now allows for short circuiting the updating of all objects. This was done to allow for asynchronous clearing of the objects table.

The camera will no longer focus on an object's center if it has one and focusing on the x,y position otherwise. Instead I just need to do Camera.target=object.center. This was done to be more efficient and to be more compatible with the newly added pan and fade states.

'exit' is now a collision class which acts simply as a solid to all entities except the player. When the player collides with it, the current level is destroyed and the next level is built by the level manager.