Added UI and Statuses classes, basic level exits
userInterface is responsible for drawing in-game text to the screen. Currently it is only damage text, but in the future it will also do dialog. Whenever damage is dealt, simple damage numbers popups appear for a brief moment, flying up and out of the damaged entity, then stopping for 1s before disappearing. I might expand this to have more variability such as numbers growing larger or having higher velocities as the damage value increases, but for now I'm happy with it.

Damage numbers can also be colored white, yellow, gray, blue, red, or green. Currently, player damage appears white, entity regular damage appears gray. Fire, Ice, and Electric damage appears red, blue, and yellow respectively. Might make health restores green. Might also need to change shades to differentiate between ally and enemy damage numbers.

The statuses class is responsible for providing status systems to entities. Currently, only 'burning' and 'frozen' statuses are implemented. If an entity is inflicted with a burn, they'll take a certain number of damage once every second for a specified duration. Burn damage knocks the enemy back weakly in a random angle. An entity inflicted with freeze will move and animate at a specified percent of their regular moveSpeed and animSpeed for a specified duration.

All basic level exits (swampWallHole, caveWallHole, and dungeonStairs) and their spawn animations are implemented. Still need a few more level exits for non-regular levels.

The Camera's curtain is now always at the camera's location, instead of the camera target location.

Refactored the takeDamage() entity and projectile method to take in a table of arguments instead of the entire reference to the damage source. This is just better practice and allows for cleaner optional arguments.

Player's bone projectile has doubled in size from 3 to 6 to better match it's shadow as well as make it easier to hit targets. Jack-O-Lantern projectile has increased in size as well.