Added entity death and projectile particles
Added particleSystem.lua which is a simple and fast version of the built in love.graphics.particleSystem module which draws particles as points instead of using image data. Easily expandable in the future, for now it just creates short, randomized particle explosions in an oval (doesn't have to be circle).

Added death explosions for all entities in the game (just missing the beholder which is yet to be added as well as all three bosses).

Added particle explosions for the jack-o-lantern and fireball projectiles as well as a small particle explosion for the blizzard projectile which is emitted each time the blizzard queries for targets.

Enemies spawns are now staggered as opposed to all spawning in at the same time.