Added Beholder entity and Orb projectile
Beholder shares behavior with standard projectile enemies, and also flies over pits. Beholder shoots 3 'orb' projectiles at a time, at a 1 radian spread.

Orb projectile is intangible, floating over both allies and enemies alike. As it travels, it will perform a check to explode every 0.5s. The shorter it's remainingTravelTime, the higher the chance to explode. Additionally, each time it bounces on a solid object, it will increase it's movement speed, up to a max of 200 units/sec.

Also added an 'activateExit() function to level exits which gets called when the player collides with a level exit.