Fixed a bug causing enemy wave to complete early
Because enemies are spawned one after each other with a 0.1s gap between each spawn, what could happen is enemies die immediately after spawning, causing the level's enemy count to reach 0 briefly before going back up when the next enemy spawns. During this brief window, the levelManager could check and see that enemyCount is 0, causing the wave/level to be completed early.

This is fixed by setting a timer for the level to only start checking for wave completion after all enemies of a wave have finished spawning.

Also increased the max speed of the orb to 300 after bouncing four times.

Also added the minus sign to the fonts