Added final boss, Witch
Witch has now been added. The Witch will be the final boss in the game, and so I plan to include a short cutscene after defeating her. Maybe some dialogue, or at least some fanfare, a wowie, and a "Happy Halloween".

The Witch has three basic attacks; a fireball, a volley of 6 icicles, and a chain lightning. These all work the same as the skeleton mages' respective attacks, except they're purple.

The Witch also has a chance to perform one of two special attacks. The chance starts out at 1/10 and increases as the Witch performs basic attacks. After 10 basic attacks the witch is guaranteed to perform a special as long as the level's enemy limit has not yet been reached.

 The first special is 'summonDemons' where she summons 1 pyreFiend, 1 beholder, 2 gnasherDemons, and 4 imps.

The second special is 'clone' where she spawns a weaker clone of herself. This clone has all of the abilities of the original, except for 'clone'. The clone's damage and health is lower than the original and it also has a weaker version of 'summonDemons' which only summons gnasherDemons and imps.

After each attack or special, the witch 50% chance to teleport around the level.