Added HUD (buttons and health)
Added the Heads Up Display

The 'buttons'  give players information about what skeleton they can summon and when they can summon it. By default, the player starts with the ability to summon skeletonWarriors, but they have to unlock skeletonArchers and skeletonMages by playing. Until then, the archer and mage icons will be greyed out. All icons will be greyed out when the player's summon ability is on cooldown, which is visualized with a small loading bar at the bottom of the button.

The health meter is a bunch of hearts, each of which represent up to 100 health, but also have sprites for a quarter, half, and threeFourths of a heart. Each time the player takes enough damage to pass a threshold of a heart piece, a small particle explosion happens at the heart which just lost a piece.

The player is now healed with the particles that explode from enemies and travel to the player, with the only exceptions being spiderEggs and tombstones. Each particle heals the player for 0.1hp, which is 10hp for typical enemies that explode into 100 particles.

There's also a skeleton counter right below the health, but I might decide to move it later.

Particle system has been reworked in order to have it's functions defined in the class, instead of in the constructor. This is more efficient and better practice.