Added most upgrade activation functions
All currently added upgrades have activation functions added except for the elemental warrior upgrades, elemental archer upgrades, and bounceShot and spreadShot for the archer. Additionally, I'd like to add more upgrades specifically for the bone category and plan to add more upgrades in general later on.

Upgrades also keeps a tally of the obtained upgrades and can correctly update it's pool to add upgrades with requirements to the pool as necessary. I'll still need to add a method of specifying and limiting how many times an upgrade can be added to the pool, if it's recurring.

Modified stats (movespeed, health, damage, etc.) of most entities to better fit in with the upgrade system.

Hud buttons now correctly display archers and mages (selected type) only when they're unlocked.