Added title screen, improved game over screen
Title screen now implemented. Just loads a test level with all enemies except for bosses, while camera pans around to follow random enemies periodically to showcase them and to add some liveliness to the screen. Title, author, and play button overlap. Play button only appears and is clickable once the fade in completes.

Increased the size of the clickable buttons used in title screen and game over screen so that they're more noticeable and better match the size of the text.

 PlayState now has two different states; 'title' and 'game'.

Fixed a bug causing crashes when a gameState not on the very top of the stack was removed. This was due to using an integer loop instead of the ipars() iterator.

Camera curtain now starts at max capacity so that the title screen can fade in more dramatically.