Added Audio class and various sound effects
Each sound effect has it's own 4 channels that it can play on. Entities will simply make a request to the Audio class to play an sfx. A random channel will then be selected to play the sfx on. This was done to keep the amount of sfx small (as opposed to giving each entity it's own sfx) while also being able to limit the amount of overlap between sfx, after all this game can get ridiculous with the amount of enemies and projectiles. At the same time, sfx aren't limited to only using a single channel so that there can be a tiny amount of overlap between the same sfx. (ex: when many bones are flying around the screen, the play will hear multiple bones hitting and bouncing off things, but there can only be at max 4 of the same sfx playing at once, to not overwhelm the player and to be more performant.

Sfx for all projectiles, there bounces, and explosions have been added as well as the default 'takeDamage' sfx (hitDefault), sfx for summon skeleton, and taking damage specifically from a 'burn' status.

Fixed a bug causing the level manager to set timers for empty functions

Fixed the specialAttacks code so that functions aren't being written to memory upon instantiation.