Added all remaining SFX, bug fixes, improvements
All SFX are finished, now all that's left is the music.

Fixed a bug causing freeze statuses still affecting the player after death. This was because the status table was cleared, but the decrease of movement and animation speed was untouched. This was fixed by restoring the entity's movement and animation speed to max in the staus:clear() function.

Player now starts with 300 hp instead of max just to better showcase the health restoring feature of killing enemies. They also definitely don't need all 500 hp at the start of the game.